Fix handling of extra in expansion attempts
If xallocx() is called with a non-zero extra, but an initial attempt to
expand an extent in-place with size+extra does not succeed, whereas the
fallback attempt of size (without extra) does succeed, this should be
reported and handled appropriately.

This fixes the case described above (initially noticed by eye, but now
verified as a bug).

Change-Id: I82c0e6eb277e53699bdb9b234fc08b442f8ac988